chore: move files for clarify
