+++ api/src/main/java/hanghae/api/adapter/MovieController.java
@@ -1,6 +1,6 @@
 package hanghae.api.adapter;
 
-import hanghae.application.dto.MoviePlayingResponse;
+import hanghae.application.dto.response.MoviePlayingResponse;
 import hanghae.application.port.MovieService;
 import hanghae.domain.type.Genre;
 import jakarta.validation.constraints.Size;
