+++ domain/src/main/java/hanghae/domain/port/MovieRepository.java
@@ -1,13 +1,10 @@
 package hanghae.domain.port;
 
-import hanghae.domain.entity.Movie;
 import hanghae.domain.type.Genre;
-import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
 import java.util.List;
 
-@Repository
 public interface MovieRepository {
     List<MovieScheduleScreenDto> findMoviesPlayingWithFilters(
             LocalDate now,
